Add using module to conn
I'm working on a plugin that could use this information, and trying to
get a hold of it via macros is a pain. I believe the code would be
better once I refactor around this change as well. I don't want to be
using macros as much as I am.